% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busselton.data.R
\docType{data}
\name{Busselton.full}
\alias{Busselton.full}
\title{Busselton dataset (full cohort)}
\format{An object of class \code{data.frame} with 1401 rows and 35 columns.}
\usage{
data(Busselton.full)
}
\description{
Busselton dataset (full cohort)
}
\keyword{datasets}
