% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcch.fit.R
\name{pcch.fit}
\alias{pcch.fit}
\title{SCAD-penalized Cox proportional hazards regression with a case-cohort design.}
\usage{
pcch.fit(time, status, covar, sub.coh, alpha = NULL, full.size = NULL,
  no.pen.covar.position = NULL, lambda.range = c(0.001, 0.5),
  nlambda = 100, threshold = 10^(-4), maxit = 200)
}
\arguments{
\item{time}{follow-up time.}

\item{status}{event indicator with 0=censored, 1=event.}

\item{covar}{covariate matrix (only numeric values are allowed).}

\item{sub.coh}{indicator of random subcohort membership with 0=not in subcohort, 1=in subcohort.}

\item{alpha}{sampling probability of the random subcohort. Required if \code{full.size} is not provided. Ignored if \code{full.size} is provided.}

\item{full.size}{full cohort size. Required if \code{alpha} is not provided.}

\item{no.pen.covar.position}{column positions of covariates that should not receive penalty. Default is NULL.}

\item{lambda.range}{range of lambda values used to create solution path. Default is [0.001, 0.5].}

\item{nlambda}{number of lambda values within \code{lambda.range}. Default is 100.}

\item{threshold}{convergence criterion. Default is 10^(-4).}

\item{maxit}{maximum number of iterations when fitting the model. Default is 200.}
}
\value{
An object of class \code{pcch.fit} containing solution path and fitted regression coefficients based on generalized cross-validation (GCV)
and BIC-type tuning parameter selection criteria.
}
\description{
\code{pcch.fit} performs SCAD-penalized Cox proportional hazards regression with a case-cohort design. The input parameters can be based on a dataset
that contains the full cohort (with the covariates of observations outside the case-cohort being missing) or a dataset that only contains the case-cohort
observations. The function produces a solution path for the given grid of tuning parameter values. It also selects the best tuning parameter and the
corresponding models based on generalized cross-validation (GCV) and BIC-type criterion. In current implementation, only SCAD penalty is available and
the number of covariates should not exceeds the number of events in the dataset.
}
\examples{
Busselton.cch=na.omit(Busselton.full)

## use dataset containing the full cohort
pcch.fit1=pcch.fit(time=Busselton.full$time, status=Busselton.full$status, covar=as.matrix(Busselton.full[,4:35]), sub.coh=Busselton.full$sub.coh.value, 
full.size=nrow(Busselton.full))

## use dataset containing only the case-cohort
pcch.fit2=pcch.fit(time=Busselton.cch$time, status=Busselton.cch$status, covar=as.matrix(Busselton.cch[,4:35]), sub.coh=Busselton.cch$sub.coh.value, 
full.size=nrow(Busselton.full))
}
\references{
Ni, A., Cai, J., and Zeng, D. (2016) "Variable selection for case-cohort studies with failure time outcome". \emph{Biometrika} \strong{103}, 3, pp. 547-562.
}
